requirements.txt
amqp==2.6.1
billiard==3.6.3.0
celery==4.4.7
Django==2.2.12
django-cors-headers==3.4.0
django-extensions==3.0.3
djangorestframework==3.9.0
importlib-metadata==1.7.0
kombu==4.6.11
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
zipp==1.2.0
